Move `SoundHelper` to `utilities.midi` package
