+++ plugin/src/main/java/com/denizenscript/denizen/scripts/commands/world/AnimateChestCommand.java
@@ -1,7 +1,7 @@
 package com.denizenscript.denizen.scripts.commands.world;
 
 import com.denizenscript.denizen.nms.NMSHandler;
-import com.denizenscript.denizen.nms.interfaces.SoundHelper;
+import com.denizenscript.denizen.utilities.midi.SoundHelper;
 import com.denizenscript.denizen.objects.LocationTag;
 import com.denizenscript.denizen.objects.PlayerTag;
 import com.denizenscript.denizen.utilities.Utilities;
