+++ plugin/src/main/java/com/denizenscript/denizen/scripts/commands/world/PlaySoundCommand.java
@@ -1,6 +1,6 @@
 package com.denizenscript.denizen.scripts.commands.world;
 
-import com.denizenscript.denizen.nms.interfaces.SoundHelper;
+import com.denizenscript.denizen.utilities.midi.SoundHelper;
 import com.denizenscript.denizen.objects.LocationTag;
 import com.denizenscript.denizen.objects.PlayerTag;
 import com.denizenscript.denizencore.exceptions.InvalidArgumentsException;
