+++ plugin/src/main/java/com/denizenscript/denizen/utilities/midi/NoteBlockReceiver.java
@@ -1,7 +1,6 @@
 package com.denizenscript.denizen.utilities.midi;
 
 import com.denizenscript.denizen.Denizen;
-import com.denizenscript.denizen.nms.interfaces.SoundHelper;
 import com.denizenscript.denizen.objects.EntityTag;
 import com.denizenscript.denizen.objects.LocationTag;
 import com.denizenscript.denizencore.utilities.debugging.Debug;
